% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sports_model.R
\name{make_sports_model}
\alias{make_sports_model}
\title{Create a new SportsModel object}
\usage{
make_sports_model(accuracy, roi, games_tested)
}
\arguments{
\item{accuracy}{Numeric accuracy (0-1)}

\item{roi}{Numeric ROI}

\item{games_tested}{Integer number of games}
}
\value{
A SportsModel object
}
\description{
Create a new SportsModel object
}
\examples{
make_sports_model(0.68, 0.12, 6000)
}
