% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edge_summary.R
\name{print.edge_summary}
\alias{print.edge_summary}
\title{Print method for edge_summary}
\usage{
\method{print}{edge_summary}(x, ...)
}
\arguments{
\item{x}{An \code{edge_summary} object.}

\item{...}{Additional arguments (unused).}
}
\description{
Print method for edge_summary
}
