% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kelly.R
\name{print.kelly_bet}
\alias{print.kelly_bet}
\title{Print method for kelly_bet}
\usage{
\method{print}{kelly_bet}(x, ...)
}
\arguments{
\item{x}{A kelly_bet object}

\item{...}{Additional arguments}
}
\description{
Print method for kelly_bet
}
