% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sports_model.R
\name{show,SportsModel-method}
\alias{show,SportsModel-method}
\title{Show method for SportsModel}
\usage{
\S4method{show}{SportsModel}(object)
}
\arguments{
\item{object}{A SportsModel object}
}
\description{
Show method for SportsModel
}
