% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sports_model.R
\docType{class}
\name{SportsModel-class}
\alias{SportsModel-class}
\title{S4 Class for Sports Model Results}
\description{
S4 Class for Sports Model Results
}
\section{Slots}{

\describe{
\item{\code{accuracy}}{Numeric. The backtested accuracy of the model.}

\item{\code{roi}}{Numeric. The Return on Investment.}

\item{\code{games_tested}}{Numeric. The number of games in the backtest.}
}}

