% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_types.R
\name{summarize_types}
\alias{summarize_types}
\title{Summarize Data Frame Types}
\usage{
summarize_types(df)
}
\arguments{
\item{df}{A data frame}
}
\value{
A data frame summarizing the columns
}
\description{
A utility function that returns a clean summary of a data frame's column types,
checking if they are S3 or S4 objects under the hood.
}
\examples{
summarize_types(mtcars)
}
