% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edge_summary.R
\name{summary.edge_summary}
\alias{summary.edge_summary}
\title{Summary method for edge_summary}
\usage{
\method{summary}{edge_summary}(object, ...)
}
\arguments{
\item{object}{An \code{edge_summary} object.}

\item{...}{Additional arguments (unused).}
}
\description{
Summary method for edge_summary
}
